Fix circular import by moving ConfigParameter to Manifest
