+++ src/steamship/invocable/config.py
@@ -1,52 +1,10 @@
 import json
-from enum import Enum
 from pathlib import Path
-from typing import Dict, Optional, Type, Union
-
-from pydantic import BaseModel, StrictBool, StrictFloat, StrictInt, StrictStr
+from typing import Dict
 
 from steamship.base.error import SteamshipError
 from steamship.base.model import CamelModel
-
-
-class ConfigParameterType(str, Enum):
-    NUMBER = "number"
-    STRING = "string"
-    BOOLEAN = "boolean"
-
-    @staticmethod
-    def from_python_type(t: Type):
-        if issubclass(t, str):
-            return ConfigParameterType.STRING
-        elif issubclass(t, bool):  # bool is a subclass of int, so must do this first!
-            return ConfigParameterType.BOOLEAN
-        elif issubclass(t, float) or issubclass(t, int):
-            return ConfigParameterType.NUMBER
-        else:
-            raise SteamshipError(f"Unknown value type in Config: {t}")
-
-    @staticmethod
-    def parse_default_value(default_value: str, t: Type):
-        if default_value is None:
-            return None
-        elif issubclass(t, str):
-            return default_value
-        elif issubclass(t, bool):  # bool is a subclass of int, so must do this first!
-            return default_value == "True"
-        elif issubclass(t, int):
-            return int(default_value)
-        elif issubclass(t, float):
-            return float(default_value)
-        else:
-            raise SteamshipError(f"Unknown value type in Config: {t}")
-
-
-class ConfigParameter(BaseModel):
-    type: ConfigParameterType
-    description: Optional[str] = None
-
-    # Use strict so that Pydantic doesn't coerce values into the first one that fits
-    default: Optional[Union[StrictStr, StrictBool, StrictFloat, StrictInt]] = None
+from steamship.data.manifest import ConfigParameter, ConfigParameterType
 
 
 class Config(CamelModel):
